[abi] move layout2 module file into directory
